define a base test client class
